+++ app/src/main/java/io/vn/nguyenduck/blocktopograph/activity/navigation/SettingFragment.java
@@ -1,5 +1,6 @@
 package io.vn.nguyenduck.blocktopograph.activity.navigation;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,13 +13,18 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import java.util.ArrayList;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 import io.vn.nguyenduck.blocktopograph.R;
 import io.vn.nguyenduck.blocktopograph.activity.subview.WorldScannerFolderView;
+import io.vn.nguyenduck.blocktopograph.setting.SettingManager;
 
 public class SettingFragment extends Fragment {
 
     private SettingAdapter adapter;
+    private final ExecutorService EXECUTOR_SERVICE = Executors.newWorkStealingPool(1);
+    private final SettingManager MANAGER = SettingManager.getInstance();
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -34,10 +40,39 @@ public class SettingFragment extends Fragment {
         return v;
     }
 
+    private void addIntoExecutor(Runnable task) {
+        if (EXECUTOR_SERVICE.isShutdown()) return;
+        EXECUTOR_SERVICE.execute(task);
+    }
+
     @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-        adapter = null;
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
+        addIntoExecutor(MANAGER::load);
+    }
+
+    @Override
+    public void onDetach() {
+        super.onDetach();
+        addIntoExecutor(MANAGER::save);
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        addIntoExecutor(MANAGER::load);
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        addIntoExecutor(MANAGER::save);
+    }
+
+    @Override
+    public void onDestroy() {
+        super.onDestroy();
+        EXECUTOR_SERVICE.shutdown();
     }
 
     private static class SettingAdapter extends BaseAdapter {